Tests/LibWeb/Text/input/CredentialManagement/FederatedCredential-ctors.html

<!DOCTYPE html>
<script src="../include.js"></script>
<script>
    test(() => {
        const init = {
            id: 'staplemachine@example.org',
            name: 'Staple Machine',
            provider: 'https://accounts.example.org',
            iconURL: 'https://example.org/images/staple.png',
            protocol: 'openidconnect',
        }
        let credential = new FederatedCredential(init);
        println(`credential instanceof FederatedCredential: ${credential instanceof FederatedCredential}`)
        println(`id: ${credential.id}`);
        println(`name: ${credential.name}`);
        println(`provider: ${credential.provider}`);
        println(`iconURL: ${credential.iconURL}`);
        println(`protocol: ${credential.protocol}`);

        // 'id' must not be empty
        try {
            new FederatedCredential({ id: '', provider: ''});
        } catch (e) {
            println(`${e.message}`);
        }
        // 'provider' must not be empty
        try {
            new FederatedCredential({ id: 'staplemachine@example.org', provider: ''});
        } catch (e) {
            println(`${e.message}`);
        }
        // 'provider' must be a valid URL
        try {
            new FederatedCredential({ id: 'staplemachine@example.org', provider: 'not a URL'});
        } catch (e) {
            println(`${e.name}: ${e.message}`);
        }
    });
</script>
