Tests/LibWeb/Text/expected/CredentialManagement/FederatedCredential-ctors.txt

credential instanceof FederatedCredential: true
id: staplemachine@example.org
name: Staple Machine
provider: https://accounts.example.org
iconURL: https://example.org/images/staple.png
protocol: openidconnect
'id' must not be empty.
'provider' must not be empty.
SyntaxError: 'provider' is not a valid url.
